fix cycle of dependency
